ready to show the group, is_live functionality with tips

+++ react-app/src/components/ProjectEdit/index.js
@@ -18,6 +18,7 @@ function Project({ everyProject }) {
   const dispatch = useDispatch();
   const history = useHistory();
   const [errors, setErrors] = useState([]);
+  const [confirmGoLive, setConfirmGoLive] = useState('')
 
   const [project, setProject] = useState("");
   const [subTitle, setSubTitle] = useState(project.sub_title);
@@ -33,7 +34,7 @@ function Project({ everyProject }) {
   );
   const [saveProjectButtonText, setSaveProjectButtonText] = useState("Save");
   const [deleteProjectButtonText, setDeleteProjectButtonText] =
-    useState("Delete");
+    useState("Delete Project");
 
   const [currentSelectedTab, setCurrentSelectedTab] = useState("basics");
 
@@ -102,14 +103,23 @@ function Project({ everyProject }) {
       if (errors.length) {
         setGoLiveErrors(errors);
       } else {
-        let projectToGoLive = {
-          ...project,
-          is_live: true,
-          created_at: new Date()
-        };
-        let projectAfterGoLive = await dispatch(editProject(projectToGoLive));
-        // console.log("*DID THIS GO LIVE? SHOULD BE YES*", projectAfterGoLive)
-        history.push(`/projects/${projectToGoLive.id}`);
+
+        if (confirmGoLive === 'Confirm') {
+
+          let projectToGoLive = {
+            ...project,
+            is_live: true,
+            created_at: new Date()
+          };
+          let projectAfterGoLive = await dispatch(editProject(projectToGoLive));
+          // console.log("*DID THIS GO LIVE? SHOULD BE YES*", projectAfterGoLive)
+          history.push(`/projects/${projectToGoLive.id}`);
+        } else {
+          setConfirmGoLive('Confirm')
+          setTimeout(() => {
+            setConfirmGoLive('')
+          }, 10000)
+        }
       }
     })();
   };
@@ -146,7 +156,6 @@ function Project({ everyProject }) {
           body: formData
         })
         const url = await response.json()
-        console.log(url, 'HERE IS YOUR URL!!!!!')
         setAwsUrl(url.url)
       })()
     }
@@ -168,7 +177,7 @@ function Project({ everyProject }) {
       await dispatch(getCategories());
       await dispatch(getSubCategories());
       await dispatch(getProjectRewards(project));
-      if (project.is_live) setCurrentSelectedTab('rewards')
+      // if (project.is_live) setCurrentSelectedTab('rewards') ** USE THIS WHEN IMPLEMENTING GO LIVE LOCKS
       await setTitle(project.title);
       await setGoal(project.funding_goal);
       await setImageUrl(project.project_image_url);
@@ -274,10 +283,17 @@ function Project({ everyProject }) {
   }
 
   const handleDelete = async (projectId) => {
-    const deletedProject = await dispatch(deleteProject(projectId));
-    if (deletedProject) {
-      history.push(`/`);
-      window.scroll(0, 0)
+    if (deleteProjectButtonText === 'Confirm') {
+      const deletedProject = await dispatch(deleteProject(projectId));
+      if (deletedProject) {
+        history.push(`/`);
+        window.scroll(0, 0)
+      }
+    } else {
+      setDeleteProjectButtonText('Confirm')
+      setTimeout(() => {
+        setDeleteProjectButtonText('Delete Project')
+      }, 10000)
     }
   };
 
@@ -572,12 +588,12 @@ function Project({ everyProject }) {
             </div>
           </form>
           <div className={styles.bottomDeleteButtonWrapper}>
-            <button
+            {/* <button
               onClick={() => handleDelete(projectId)}
               className={styles.bottomDeleteButton}
             >
               {deleteProjectButtonText}
-            </button>
+            </button> */}
           </div>
         </div>
       );
@@ -623,21 +639,49 @@ function Project({ everyProject }) {
               </button>
             </div>
           </div>
-          <div className={styles.topRight}>
-            <div className={styles.buttonDiv}>
-              <Link to="/">
-                <button className={styles.cancelButton}>Cancel</button>
-              </Link>
-            </div>
+          <div style={{ display: 'flex', alignItems: 'center', flexDirection: 'column', justifyContent: 'center' }}>
 
-            <div className={styles.buttonDiv}>
-              <button
-                onClick={handleGoLiveButton}
-                className={styles.saveButton}
-              >
-                Go Live
-              </button>
+
+            {deleteProjectButtonText === 'Confirm' &&
+              <div className={styles.impression} style={{ marginBottom: '7px', display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
+                <span>
+                  <i class="fas fa-lightbulb"></i>
+                </span>
+                <span style={{ marginLeft: '3px' }}>
+                  {" "}
+                  Are you sure? This action will permanently delete your project. Click to confirm. {" "}
+                </span>
+              </div>
+            }
+            {confirmGoLive === "Confirm" &&
+              <div className={styles.impression} style={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
+                <span>
+                  <i class="fas fa-lightbulb"></i>
+                </span>
+                <span style={{ marginLeft: '3px' }}>
+                  {" "}
+                  Ready to go live? You will no longer be able to edit project Basics/Story. You will maintain access to Rewards. Click to confirm! {" "}
+                </span>
+              </div>
+            }
+          </div>
+          <div className={styles.topRight}>
+            <div className={styles.buttonDiv} style={project.is_live ? { width: '100%' } : null}>
+              {/* <Link to="/"> */}
+              <button className={styles.cancelButton}
+                onClick={() => handleDelete(projectId)}>{deleteProjectButtonText}</button>
+              {/* </Link> */}
             </div>
+            {!project?.is_live &&
+              <div className={styles.buttonDiv}>
+                <button
+                  onClick={handleGoLiveButton}
+                  className={styles.saveButton}
+                >
+                  {confirmGoLive ? 'Confirm' : "Go Live"}
+                </button>
+              </div>
+            }
           </div>
         </div>
         <div>
@@ -701,6 +745,20 @@ function Project({ everyProject }) {
             </div>
           </div>
         </div>
+        {!project?.is_live &&
+          <div className={styles.noteSection}>
+
+            <div className={styles.impression}>
+              <span>
+                <i class="fas fa-lightbulb"></i>
+              </span>
+              <span>
+                {" "}
+                Note: Your project is not yet live. Your project will not be displayed publicly. {" "}
+              </span>
+            </div>
+          </div>
+        }
       </div>
       {dashBoardContent()}
     </div>
